# Performance improvements to BinaryMinMaxHeap